requirements_cpu.txt
absl-py==2.0.0
pandas==2.0.3
google-cloud-storage==2.13.0
scikit-learn==1.3.0
transformers==4.31.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1